[ALPS03194775] M4U:fix m4u code defect
[Detail] fix m4u get pre info pa type

MTK-Commit-Id: 49953b4657a7b211814850ac8bc72e050fb91593

CR-Id:ALPS03194775
Feature:others
(cherry picked from commit 9ff99d4be997de294bf3e93b5544d92c330b1100)

+++ drivers/misc/mediatek/m4u/2.0/m4u_pgtable.c
@@ -141,7 +141,7 @@ int m4u_get_pte_info(struct m4u_domain *domain, unsigned int mva, m4u_pte_info_t
 {
 	imu_pgd_t *pgd;
 	imu_pte_t *pte = NULL;
-	unsigned long pa = 0;
+	unsigned long long pa = 0;
 	unsigned int size;
 	int valid = 1;
